Auth now used every service and Few request method changed

+++ cart/src/main/java/org/ewebshop/cart/CartService.java
@@ -3,6 +3,7 @@ package org.ewebshop.cart;
 import lombok.AllArgsConstructor;
 import org.ewebshop.commons.clients.order.OrderClient;
 import org.ewebshop.commons.clients.product.ProductClient;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.EntityNotFoundException;
@@ -18,7 +19,7 @@ public class CartService {
     private final ProductClient productClient;
 
     public List<Cart> getAll(int orderId) {
-        if (orderClient.existsCheck(orderId)){
+        if (orderClient.existsCheck(orderId, "Bearer " + (String) SecurityContextHolder.getContext().getAuthentication().getCredentials())){
             return cartRepository.findAllByOrderId(orderId);
         } else {
             throw new EntityNotFoundException("Order is not existing");
@@ -30,19 +31,19 @@ public class CartService {
     }
 
     public void addCartItem(int productId, int orderId, int quantity) {
-        if (orderClient.existsCheck(orderId)){
+        if (orderClient.existsCheck(orderId, "Bearer " +  (String) SecurityContextHolder.getContext().getAuthentication().getCredentials())){
             if (productClient.existsCheck(productId)) {
                 Optional<Cart> cart = cartRepository.findById(new CartId(productId,orderId));
                 if (cart.isPresent()) {
                     Cart cartEntity = cart.get();
                     cartEntity.setQuantity(cartEntity.getQuantity() + quantity);
                     //csökkentem a product mennyiséget a raktárban
-                    productClient.decreaseQuantity(productId, quantity);
+                    productClient.decreaseQuantity(productId, quantity,"Bearer " +  (String) SecurityContextHolder.getContext().getAuthentication().getCredentials());
                     cartRepository.save(cartEntity);
                 } else {
                     Cart newCartEntity = new Cart(productId,orderId,quantity);
                     //csökkentem a product mennyiséget a raktárban
-                    productClient.decreaseQuantity(productId, quantity);
+                    productClient.decreaseQuantity(productId, quantity,"Bearer " +  (String) SecurityContextHolder.getContext().getAuthentication().getCredentials());
                     cartRepository.save(newCartEntity);
                 }
             } else {
@@ -54,20 +55,25 @@ public class CartService {
     }
 
     public void removeCartItem(int productId, int orderId, int quantity) {
-        Optional<Cart> cart = cartRepository.findById(new CartId(productId,orderId));
-        if (cart.isPresent()) {
-            Cart cartEntity = cart.get();
-            cartEntity.setQuantity(cartEntity.getQuantity() - quantity);
-            if (cartEntity.getQuantity() <= 0) {
-                cartRepository.delete(cartEntity);
-            } else {
-                cartRepository.save(cartEntity);
-            }
-            productClient.increaseQuantity(productId, quantity);
+        if (orderClient.existsCheck(orderId , "Bearer " +  (String) SecurityContextHolder.getContext().getAuthentication().getCredentials())) {
+            Optional<Cart> cart = cartRepository.findById(new CartId(productId,orderId));
+            if (cart.isPresent()) {
+                Cart cartEntity = cart.get();
+                cartEntity.setQuantity(cartEntity.getQuantity() - quantity);
+                if (cartEntity.getQuantity() <= 0) {
+                    cartRepository.delete(cartEntity);
+                } else {
+                    cartRepository.save(cartEntity);
+                }
+                productClient.increaseQuantity(productId, quantity,"Bearer " +  (String) SecurityContextHolder.getContext().getAuthentication().getCredentials());
 
+            } else {
+                throw new EntityNotFoundException("CartItem is not existing");
+            }
         } else {
-            throw new EntityNotFoundException("CartItem is not existing");
+            throw new EntityNotFoundException("Order is not existing");
         }
+
     }
 
 }
